skin_ai_assistant/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
pydantic
onnxruntime
opencv-python
numpy
pillow
torch
torchvision
streamlit
requests
pytest
httpx
